make order of lines in top_level.txt deterministic
without this, every build produces a different result, as the lines are
ordered randomly.

This makes builds reproducible.
If you build a package two times, you get the same result, as you would
expect. Basically the same thing was already done for requirements and
entry_points.txt

It is useful e.g. if you want to rebuild a package that you downloaded,
to ensure that no-one tampered with it, or even if you just want to
examine the differences between two builds/versions of one package.