sdist command: fix case insensitivity when adding some files to filelist
This should fix the problem in Bitbucket issue #100.  It gives the same behavior
for inclusion of default files (README*, etc.) on Windows as Linux.

BACKWARDS INCOMPATABILITY:

This may result in a backwards incompatible change for users on a
case insensitive file system.  If they were relying on some files getting
included in their distribution due to setuptools defaults, and their files do not
have the same case as the files being looked for in setuptools, those files
will no longer be included in the package.

For example, if a package had a file:

readme.rst

Previous to this commit, that file would have been included in the distribution
as:

README.rst

But it will now no longer be included at all.  To get the file included in
the package, it can be added to the package's MANIFEST.in file:

include readme.rst

Files affected by this change will have a case variant of the files or
patterns listed below:

README
README.txt
README.rst
setup.py (or whatever your setuptools script is named)
setup.cfg
test/test*.py