egg_info: Search egg-base for files to add to the manifest Before this commit, this:
  $ mkdir -p /tmp/xyz/{home,lib,scripts,data,egg}
  $ cat >/tmp/xyz/home/.pydistutils.cfg <<EOF
  > [egg_info]
  > egg-base = /tmp/xyz/egg
  > EOF
  $ export PYTHONPATH=/tmp/xyz/lib
  $ export HOME=/tmp/xyz/home
  $ setup.py install --home=/tmp/xyz/home --install-lib=/tmp/xyz/lib \
  >   --install-scripts=/tmp/xyz/scripts --install-data=/tmp/xyz/data

drops a lot of metadata, installing only SOURCES.txt and zip-safe
under EGG-INFO.  The problem is that the metadata files are written to
egg-base, but egg-base is not searched when creating the manifest
because it's outside of the current directory.  Work around this by
explicitly searching egg-base with distutils.filelist.findall (which
is really the version monkeypatched in by setuptools/__init__.py).

Since findall records relative paths, prefix the returned paths with
egg-base, so the include_pattern looking for the absolute
ei_cmd.egg_info will match them.