Sort __path__ entries for namespace packages according to their order in sys.path.  This ensures that lookups in __path__ will be the same as sys.path resolution.
This also adds a replace argument to Distribution.insert_on meant
to be used with the replace argumen to WorkingSet.add.  This ensures
that new sys.path entries added via WorkingSet.add are inserted at
the beginning, rather than appended to the end.  This is necessary
for consistency with the above change, and kind of makes more sense
anyways.  This means that if a Distribution is added to a WorkingSet,
that replaces a different version of that Distribution, the new version
of that Distribution will have its location first on sys.path.